+++ sw-sysemu/insns/tensors.cpp
@@ -1203,11 +1203,14 @@ void tensor_store_execute(Hart& cpu)
     }
 
 #ifdef SYS_EMU
+    bool tstoreCheck = SYS_EMU_PTR->get_tstore_check();
+    auto& tstoreChecker = SYS_EMU_PTR->get_tstore_checker();
+
     // After all the checks are done, report the TStore access, notice that TStores that
     // are partially done due a tensor error won't be captured...
-    if (SYS_EMU_PTR->get_tstore_check() && (coop > 1)) {
-        SYS_EMU_PTR->get_tstore_checker().execute(hart_index(cpu), addr, stride, coop, rows, cols);
-        SYS_EMU_PTR->get_tstore_checker().check_and_drain(hart_index(cpu));
+    if ((tstoreCheck) && (coop > 1)) {
+        tstoreChecker.execute(hart_index(cpu), addr, stride, coop, rows, cols);
+        tstoreChecker.check_and_drain(hart_index(cpu));
     }
 #endif